requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
decorator==4.4.2
flatbuffers==1.12
gast==0.4.0
google-api-core==1.34.0
google-api-python-client==2.4.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.22.0
idna==3.4
imageio==2.26.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
joblib==1.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.3.2
mhyt==3.5.5
moviepy==1.0.3
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.5.3
Pillow==9.4.0
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scenedetect==0.5.5
scikit-learn==0.23.2
scipy==1.10.1
seaborn==0.11.0
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
tkVideo==0.1
tqdm==4.65.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.12.1
youtube-dl==2021.12.17
zipp==3.15.0
